Add popcorn.c and Combine Functionality
Fleshed out and finally merged the functionality of `consts.c` and
`display.c`, now with the primary file `popcorn.c` to be a parent to
both. Popcorn will now read a config file, discern the media directory,
scan that directory for files, populate the state array with the files,
and then display the ncurses ui, which is a pretty good start.