Change Config Storage, Change Layout, & Add Save Feature
Chaged the way the configuration is stored, now in a struct that lives
in `popcorn.c` intead of a bunch of gross global variables. Changed the
layout to a new stacked format, which I think fits the content better.
Also added functionality for saving and loading the media array, which is
a sort of stand in for a sqlite database that will probably be
implemented in the future.