Add Edit Functionality
Added new functionality to open a temporary file in $EDITOR containing
information on the selected title. This can then be edited, and once
saved will be changed in the array. Saving and loading from the store
file is now more important because of editing, so warnings have been
added in case an action will override saved changes. `display.c` and
it's header have been renamed to `ui` since the file both display and
handles input. The configuration is now stored in a struct that lives in
main(), which is much cleaner than the previous global variables.