Add Info Display and Change Directory Scanning
Added more information to the info area on the bottom of the screen. Now
the summary, genre, year, runtime, and seen data is shown. Added genre
and runtime data to the media struct. Added a new request system when
scanning a directory to ask the user if they want to change the metadata
(most likely will since API for movie info isn't integrated yet so
everything is set to a default value). There is a bottom bar now which
should hold controls.