Add OMDB Support and Video Player Functionality
Added support for getting metadata from the [Open Movie
Database](https://www.omdbapi.com/) if an api key is provided in the
config file. This can be done either at the point the title is added
into popcorn or by pressing 'o' on the title in the user interface.
Also added the functionality to open a video player by pressing enter on
a title.