Fix Media Player Waiting and Improve Resizing
Fixed the problem where opening a movie would cause the curses ui to
glitch out and sent all input to mpv instead. Now, popcorn won't attempt
to wait for your media player to finish since that was kind of a dumb
idea in the first place. Also improved resizing, in that popcorn now
actually attempts to do it when the window is resized. It's still pretty
janky, and doesn't work when trying to resize the window larger than it
was when it initially started up, but it's an improvement over what it
was doing before.